requirements.txt
torch==2.2.0
transformers==4.37.2
pillow==10.2.0
numpy==1.26.4
tqdm==4.66.2
scikit-learn==1.4.2
numpy==1.26.4
tqdm==4.66.2
scikit-learn==1.4.2